fullbright fix & Music player + other

+++ net/minecraft/client/gui/FontRenderer.java
@@ -374,15 +374,15 @@ public class FontRenderer implements IResourceManagerReloadListener
     /**
      * Draws the specified string.
      */
-    public int drawString(String text, int x, int y, int color)
+    public int drawString(String text, double x, double y, int color)
     {
-        return !this.enabled ? 0 : this.drawString(text, (float)x, (float)y, color, false);
+        return !this.enabled ? 0 : this.drawString(text, x, y, color, false);
     }
 
     /**
      * Draws the specified string.
      */
-    public int drawString(String text, float x, float y, int color, boolean dropShadow)
+    public int drawString(String text, double x, double y, int color, boolean dropShadow)
     {
         this.enableAlpha();
         this.resetStyles();
@@ -616,7 +616,7 @@ public class FontRenderer implements IResourceManagerReloadListener
     /**
      * Render single line string by setting GL color, current (posX,posY), and calling renderStringAtPos()
      */
-    private int renderString(String text, float x, float y, int color, boolean dropShadow)
+    private int renderString(String text, double x, double y, int color, boolean dropShadow)
     {
         if (text == null)
         {
@@ -644,8 +644,8 @@ public class FontRenderer implements IResourceManagerReloadListener
             this.green = (float)(color & 255) / 255.0F;
             this.alpha = (float)(color >> 24 & 255) / 255.0F;
             this.setColor(this.red, this.blue, this.green, this.alpha);
-            this.posX = x;
-            this.posY = y;
+            this.posX = (float)x;
+            this.posY = (float)y;
             this.renderStringAtPos(text, dropShadow);
             return (int)this.posX;
         }
